reports: Update to use report_prefix
This commit removes the report_filename config parameter and adds a
report_prefix parameter.  AFAICT the current code suggests that it will
write all host data to the same file.  However the code writes the old
data to the prom file giving all values -1 which is incorrect

I looked at why this was and i can't find much information on the
original issue[1] and PR[2].  further to this the current implementation
is very racy if one sets REPORT_FILENAME

This also updates the clean stale function toi only act on files with
the report_prefix and avoid deleting files which may have been placed
there by none puppetserver process

[1]https://github.com/voxpupuli/puppet-prometheus_reporter/pull/5
[2]https://github.com/voxpupuli/puppet-prometheus_reporter/pull/7